adding font size by param.

+++ plotex.sh
@@ -1,12 +1,29 @@
 #!/bin/sh
 
+ARGC=$#
+fontsize='\\tiny'
 
-gnuplot   -e "set term epslatex color solid  font ',8' header '\\tiny'; set output '${PLOTEXDIR}src/fig1.tex'" $1
-jobname="${1%.*}"
+if [ $ARGC -eq 0 ]
+then
+	echo 'filename of gnuplot script?'
+fi
 
-#pdflatex  --output-directory=out --jobname=$jobname src/plot.tex
-pdflatex  -shell-escape --jobname=$jobname ${PLOTEXDIR}src/plot.tex 
-rm ${jobname}.aux
-rm ${jobname}.log
+if [ $ARGC -gt 0 ]
+then
+	
+	if [ $ARGC -gt 1 ]
+	then
+		fontsize=$2
+		gnuplot   -e "set term epslatex color solid  font ',8' header '\\${fontsize}'; set output '${PLOTEXDIR}src/fig1.tex'" $1
+	else
+		gnuplot   -e "set term epslatex color solid  font ',8' header '\\tiny'; set output '${PLOTEXDIR}src/fig1.tex'" $1
 
-evince ${jobname}.pdf &
+	fi
+		
+	jobname="${1%.*}"
+	pdflatex  -shell-escape --jobname=$jobname ${PLOTEXDIR}src/plot.tex 
+	rm ${jobname}.aux
+	rm ${jobname}.log
+	evince ${jobname}.pdf &
+
+fi
